hall_and_book_validation

+++ app/Http/Controllers/BookingController.php
@@ -2,9 +2,13 @@
 
 namespace App\Http\Controllers;
 
+use App\Http\Requests\Booking\CreateBookRrequest;
+use App\Http\Requests\Booking\UpdateBookIdRrequest;
 use App\Models\Booking;
 use Illuminate\Http\Request;
 use Illuminate\Support\Facades\DB;
+use Illuminate\Support\Facades\Validator;
+
 
 class BookingController extends Controller
 {
@@ -24,22 +28,49 @@ class BookingController extends Controller
      * @param  \Illuminate\Http\Request  $request
      * @return \Illuminate\Http\Response
      */
-    public function store(Request $request)
+    public function store(CreateBookRrequest $request)
     {
-        $booking = Booking::create($request->all()); 
-        return $booking;
+        $data = $request->all();
+        $concatenatedData = $request->get('hall_num_id') . '-' . $request->get('start_time_booking') . '-'. $request->get('end_time_booking');
+        
+        $validator = Validator::make(['concatenated_data' => $concatenatedData], [
+            'concatenated_data' => 'unique:bookings,concatenated_data',
+        ]);
+    
+        if ($validator->fails()) {
+            return response()->json(['message' => 'There is an employee that booked the same hall in the same time.'], 422);
+        }
+
+        DB::table('bookings')->insert([
+            'employee_num_id' => $request->get('employee_num_id'),
+            'hall_num_id' => $request->get('hall_num_id'),
+            'start_time_booking' => $request->get('start_time_booking'),
+            'end_time_booking' => $request->get('end_time_booking'),
+            'booking_day' => $request->get('booking_day'),
+            'concatenated_data' => $concatenatedData,
+            ]);
+    
+        return $data;
     }
 
     /**
      * Display the specified resource.
      *
-     * @param  int  $id
-     *  @param  int  $id
+     * @param  int  $employee_num_id
+     *  @param  int  $hall_num_id
      * @return \Illuminate\Http\Response
      */
-    public function show($id1,$id2)
+    public function show($employee_num_id,$hall_num_id)
     {
-        return Booking::where('employee_num_id',$id1)->where('hall_num_id',$id2)->firstOrFail(); 
+        if (filled($employee_num_id) && filled($hall_num_id) && is_numeric($employee_num_id) && is_numeric($hall_num_id)) {
+            $bookings = Booking::where('employee_num_id', $employee_num_id)
+                                ->where('hall_num_id', $hall_num_id)
+                                ->get();
+            return response()->json($bookings);
+        }
+         else {
+            return response()->json(['message' => 'Invalid input.'], 400);
+        }
 
     }
 
@@ -47,30 +78,49 @@ class BookingController extends Controller
      * Update the specified resource in storage.
      *
      * @param  \Illuminate\Http\Request  $request
-     * @param  int  $id1
-     *  @param  int  $id2
+     * @param  int  $employee_num_id
+     *  @param  int  $hall_num_id
+     *  @param  int  $id
      * @return \Illuminate\Http\Response
      */
-    public function update(Request $request,$id1,$id2)
+    public function update(UpdateBookIdRrequest $request,$id,$employee_num_id,$hall_num_id)
     {
+        if (filled($id) && filled($employee_num_id) && filled($hall_num_id) && is_numeric($employee_num_id) && is_numeric($hall_num_id) && is_numeric($id)) {
+            $bookings =DB::table('bookings')
+            ->where('id',$id)
+            ->where('employee_num_id',$employee_num_id)
+            ->where('hall_num_id',$hall_num_id)
+            ->update($request->all());
+            return response()->json($bookings);
+            } 
+        else {
+            return response()->json(['message' => 'Invalid input.'], 400);
+        }
         
-        return DB::table('bookings')
-        ->where('employee_num_id',$id1)
-        ->where('hall_num_id',$id2)
-        ->update($request->all());
+        
     }
 
     /**
      * Remove the specified resource from storage.
      *
-     * @param  int  $id1
-     * * @param  int  $id2
+     * @param  int  $employee_num_id
+     * * @param  int  $hall_num_id
+     *  @param  int  $id
      * @return \Illuminate\Http\Response
      */
-    public function destroy($id1,$id2)
+    public function destroy($id,$employee_num_id,$hall_num_id)
     {
-        $booking = Booking::where('employee_num_id',$id1)->where('hall_num_id',$id2);
-        $booking->delete();
-        return response('',204);
+        if (filled($id) && filled($employee_num_id) && filled($hall_num_id) && is_numeric($employee_num_id) && is_numeric($hall_num_id) && is_numeric($id)) {
+            $booking =DB::table('bookings')
+            ->where('id',$id)
+            ->where('employee_num_id',$employee_num_id)
+            ->where('hall_num_id',$hall_num_id);
+            $booking->delete();
+            return response('',204);
+            } 
+        else {
+            return response()->json(['message' => 'Invalid input.'], 400);
+        }
+        
     }
 }
